Change the implementation in order to follow the new convections with dotnet 2.1.x
Now the EnsureSeedData method is not an extension method of IApplicationBuilder.
The user has to create an scope an pass it to the method.

The call to the EnsureSeedData has to be done in the Main method that can be found
in the Program.cs file of the dotnet core projects.